tests/submission/Microsoft/forms/formenctype.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>formenctype on button element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="formenctype support on elements" />
    </head>
    <body>
        <script src="/resources/testharness.js" ></script>
        <script src="/resources/testharnessreport.js" ></script>

        <div id=log></div>
        <button style="display: none" type="submit" formenctype="text/plain" >Submit</button>
        <input style="display: none" type="submit" formenctype="text/plain" value="submit"></input>
        <input style="display: none" type="submit" value="submit"></input>
        
        <script type="text/javascript">
        
            test(function() {assert_equals(document.getElementsByTagName("button")[0].formEnctype, "text/plain")}, "formEnctype support on button");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].formEnctype, "text/plain")}, "formEnctype support on input");
        
            var testElem = document.getElementsByTagName("input")[0];
            testElem.formEnctype = "multipart/form-data";

            test(function() {assert_equals(document.getElementsByTagName("input")[0].formEnctype, "multipart/form-data")}, "formEnctype on input reflects correct value after being updated by the DOM");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].getAttribute("formenctype"), "multipart/form-data")}, "formEnctype value is correct using getAttribute");
                
            test(function() {assert_equals(document.getElementsByTagName("input")[1].formEnctype, "application/x-www-form-urlencoded")}, "formEnctype defaults to application/x-www-form-urlencoded");
        </script>
    </body>
</html>
